feat: implement [slog.LogValuer] for [hcloud.Error]
Related to #826

Instead of logging only the error String() as message, make a group with the error message and the error details when they are available.

```rp-commits
feat: implement [slog.LogValuer] for [hcloud.Error]
feat(exp): remove errutil package
```
